Server/pyproject.toml
[project]
name = "mal-2-yt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "fastapi[standard]>=0.116.1",
    "itsdangerous>=2.2.0",
    "mal-api-py>=0.10.1",
    "python-dotenv>=1.1.1",
]
